+++ surepy/data/track.py
@@ -45,9 +45,9 @@ def link_locdata(locdata, search_range=40, memory=0, **kwargs):
     """
     df = link_df(locdata.data, search_range=search_range, memory=memory, pos_columns=locdata.coordinate_labels,
                     t_column='Frame', **kwargs)
-    df.reset_index(inplace=True)
-    df.rename(columns={'index': 'Index', 'particle': 'Track'}, inplace=True)
-    return df[['Index', 'Track']]
+    return_series = df['particle']
+    return_series.name = 'Track'
+    return return_series
 
 
 def track(locdata, search_range=40, memory=0, **kwargs):
@@ -77,11 +77,9 @@ def track(locdata, search_range=40, memory=0, **kwargs):
     """
     parameter = locals()
 
-    df = link_locdata(locdata,search_range, memory, **kwargs)
-
-    grouped = df.groupby('Track')
-
-    selections = [LocData.from_selection(locdata=locdata, indices=group['Index'].values) for _, group in grouped]
+    track_series = link_locdata(locdata, search_range, memory, **kwargs)
+    grouped = track_series.groupby(track_series)
+    selections = [LocData.from_selection(locdata=locdata, indices=group.index.values) for _, group in grouped]
     collection = LocData.from_collection(selections)
 
     # metadata