+++ surepy/tests/data/test_track.py
@@ -3,7 +3,7 @@ import numpy as np
 import pandas as pd
 
 from surepy import LocData
-from surepy.data.track import track
+from surepy.data.track import link_locdata, track
 
 
 @pytest.fixture()
@@ -16,6 +16,12 @@ def locdata_simple():
 
     }
     return LocData(dataframe=pd.DataFrame.from_dict(dict))
+
+def test_link_locdata(locdata_simple):
+    track_series = link_locdata(locdata_simple, search_range=5, memory=0)
+    assert (len(track_series) == 8)
+    assert(track_series.name=='Track')
+
 
 def test_track(locdata_simple):
     locdata_new = track(locdata_simple, search_range=5)